🔧 chore(helpers.test.js): add test for slug helper function
🔧 chore(app.js): import slug and vLog helper functions from helpers module

🔧 chore(helpers.js): add slug and vLog helper functions

The changes were made to add a test for the slug helper function in the helpers.test.js file. Additionally, the slug and vLog helper functions were added to the helpers.js file. The app.js file was modified to import the slug and vLog helper functions from the helpers module.
